Renaming the model
