+++ server/database.ts
@@ -88,7 +88,7 @@ import WaitlistRegistrationFactory, {
 } from './models/waitlist_registration';
 import WebhookFactory, { WebhookModelStatic } from './models/webhook';
 import NotificationsReadFactory, { NotificationsReadModelStatic } from './models/notifications_read';
-import IpfsPinsFactory, { IpfsPinsModelStatic } from './models/ipfspins';
+import IpfsPinsFactory, { IpfsPinsModelStatic } from './models/ipfs_pins';
 
 export type Models = {
   Address: AddressModelStatic;
